Fixed some sonartype reportings. Removed xsddoc as it just get outdated on the apache site.
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@797459 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/main/java/org/apache/camel/builder/DefaultErrorHandlerBuilder.java
@@ -71,7 +71,7 @@ public class DefaultErrorHandlerBuilder extends ErrorHandlerBuilderSupport {
         return this;
     }
 
-    public DefaultErrorHandlerBuilder collisionAvoidancePercent(short collisionAvoidancePercent) {
+    public DefaultErrorHandlerBuilder collisionAvoidancePercent(double collisionAvoidancePercent) {
         getRedeliveryPolicy().collisionAvoidancePercent(collisionAvoidancePercent);
         return this;
     }
