+++ camel-core/src/main/java/org/apache/camel/model/OnExceptionDefinition.java
@@ -286,7 +286,7 @@ public class OnExceptionDefinition extends ProcessorDefinition<ProcessorDefiniti
      * @param collisionAvoidancePercent  the percentage
      * @return the builder
      */
-    public OnExceptionDefinition collisionAvoidancePercent(short collisionAvoidancePercent) {
+    public OnExceptionDefinition collisionAvoidancePercent(double collisionAvoidancePercent) {
         getOrCreateRedeliveryPolicy().collisionAvoidancePercent(collisionAvoidancePercent);
         return this;
     }
